@@ -1,31 +1,33 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
+
+plugins {
+    id 'com.android.application'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-parcelize'
+}
 
 
 android {
     signingConfigs {
         debugincloud {
             keyAlias 'debugcloud'
-            storeFile file('/Users/williamwu/Programming/IOTDevice/cloud')
+            storeFile file('/home/william/AndroidStudioProjects/IOT-Loopback-Android/IOTDevice/cloud')
             keyPassword '1qaz@WSX'
             storePassword '1qaz@WSX'
         }
         clouddebug {
             keyAlias 'cloud'
-            storeFile file('/Users/williamwu/Programming/IOTDevice/clouddebug.keystore')
+            storeFile file('/home/william/AndroidStudioProjects/IOT-Loopback-Android/clouddebug.keystore')
             keyPassword '1qaz@WSX'
             storePassword '1qaz@WSX'
         }
     }
-    compileSdkVersion 26
+    compileSdkVersion 34
     defaultConfig {
         applicationId "iotdevice.com.iot_device"
         minSdkVersion 21
-        targetSdkVersion 26
-        versionCode 10
-        versionName "1.8.1"
+        targetSdk 34
+        versionCode 25
+        versionName "2.2.5"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
