@@ -5,6 +5,7 @@ import android.accounts.AccountManager
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import iotdevice.com.iotDevice.common.DialogUtils
 import iotdevice.com.iotDevice.common.IOTPreference
 import iotdevice.com.iotDevice.home.HomeActivity
@@ -15,15 +16,13 @@ import iotdevice.com.iotDevice.model.CustomerModel
 import iotdevice.com.iotDevice.register.RegisterActivity
 import iotdevice.com.iotDevice.resetpassword.ResetPasswordActivity
 import iotdevice.com.iot_device.R
-import kotlinx.android.synthetic.main.activity_login.*
-import org.jetbrains.anko.AnkoLogger
-import org.jetbrains.anko.info
-import org.jetbrains.anko.startActivity
+import iotdevice.com.iot_device.databinding.ActivityLoginBinding
 
 
-class LoginActivity : AccountAuthenticatorActivity(), AnkoLogger, TokenManager.LoginListener {
+class LoginActivity : AccountAuthenticatorActivity(), TokenManager.LoginListener {
 
     var mAccountManager: AccountManager? = null
+    lateinit var binding: ActivityLoginBinding
 
     companion object {
         const val ARG_ACCOUNT_TYPE = "accountType"
@@ -31,24 +30,27 @@ class LoginActivity : AccountAuthenticatorActivity(), AnkoLogger, TokenManager.L
         const val ARG_IS_ADDING_NEW_ACCOUNT = "isAddingNewAccount"
         const val PARAM_USER_PASSWORD = "password"
         const val KEY_IS_GLOBAL_PROFILE = "IS_GLOBAL_PROFILE"
+        const val tag = "LoginActivity"
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
+        binding = ActivityLoginBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         mAccountManager = AccountManager.get(this)
 //        val loginViewModel = ViewModelProviders.of(this).get(LoginViewModel::class.java)
 
-        loginBtn.setOnClickListener({ _ ->
+        binding.loginBtn.setOnClickListener({ _ ->
             checkUserInput()
         })
 
-        registerTextView.setOnClickListener({ _ ->
+        binding.registerTextView.setOnClickListener({ _ ->
             startActivity(RegisterActivityIntent())
         })
 
-        forgotPasswordTextView.setOnClickListener({ _ ->
-            startActivity<ResetPasswordActivity>()
+        binding.forgotPasswordTextView.setOnClickListener({ _ ->
+            val intent = Intent(applicationContext,ResetPasswordActivity::class.java)
+            startActivity(intent)
         })
 
         val intent = Intent(this, HomeActivity::class.java)
@@ -68,11 +70,11 @@ class LoginActivity : AccountAuthenticatorActivity(), AnkoLogger, TokenManager.L
 
     override fun onLoginComplete(result: CustomerModel) {
         val authToken = result.id
-        val accountName = loginUserName.text.toString()
-        val accountPassword = password_Edit.text.toString()
+        val accountName = binding.loginUserName.text.toString()
+        val accountPassword = binding.passwordEdit.text.toString()
         val account = Account(accountName, AuthUtil.ACCOUNT_TYPE_NAME)
 
-        info("authToken : $authToken")
+        Log.i(tag, "authToken : $authToken")
         if (intent.getBooleanExtra(ARG_IS_ADDING_NEW_ACCOUNT, true)) {
             // Creating the account on the device and setting the auth token we got
             // (Not setting the auth token will cause another call to the server to authenticate the user)
@@ -82,21 +84,23 @@ class LoginActivity : AccountAuthenticatorActivity(), AnkoLogger, TokenManager.L
         } else {
             mAccountManager?.setPassword(account, accountPassword)
         }
-        val intent = Intent(this, HomeActivity::class.java)
-//        intent.putExtra(AccountManager.KEY_ACCOUNT_NAME, accountName)
-//        intent.putExtra(AccountManager.KEY_ACCOUNT_TYPE, AuthUtil.AUTH_TOKEN_TYPE_NAME)
-        intent.putExtra(AccountManager.KEY_AUTHTOKEN, authToken)
-        intent.putExtra(AccountManager.KEY_PASSWORD, accountPassword)
-        setAccountAuthenticatorResult(intent.extras)
-        setResult(RESULT_OK, intent)
+        val intent = Intent(this, HomeActivity::class.java).apply {
+            putExtra(AccountManager.KEY_AUTHTOKEN, authToken)
+            putExtra(AccountManager.KEY_PASSWORD, accountPassword)
+            setAccountAuthenticatorResult(extras!!)
+            setResult(RESULT_OK, intent)
+        }
 
         IOTPreference.saveUserName(result.username)
         startActivity(intent)
+//        intent.putExtra(AccountManager.KEY_ACCOUNT_NAME, accountName)
+//        intent.putExtra(AccountManager.KEY_ACCOUNT_TYPE, AuthUtil.AUTH_TOKEN_TYPE_NAME)
+
         finish()
     }
 
     override fun onLoginError(err: Throwable) {
-        info("login error $err")
+        Log.i(tag, "login error $err")
 //        DialogUtils.createAlertDialog( this, getString(R.string.login_title), msg = err.message)
         DialogUtils.createAlertDialog( this, getString(R.string.login_title), msg = getString(R.string.login_fail))
 
@@ -108,10 +112,10 @@ class LoginActivity : AccountAuthenticatorActivity(), AnkoLogger, TokenManager.L
     }
 
     private fun checkUserInput() {
-        if (loginUserName.text.toString().isNotEmpty() &&
-                password_Edit.text.toString().isNotEmpty()) {
-            TokenManager.performLoginRequest(loginUserName.text.toString(),
-                    password_Edit.text.toString(), this)
+        if (binding.loginUserName.text.toString().isNotEmpty() &&
+                binding.passwordEdit.text.toString().isNotEmpty()) {
+            TokenManager.performLoginRequest(binding.loginUserName.text.toString(),
+                    binding.passwordEdit.text.toString(), this)
         } else {
             DialogUtils.createAlertDialog( this, getString(R.string.login_title), msg = resources.getString(R.string.loginin_field_not_completed))
         }
