@@ -1,16 +1,16 @@
 package iotdevice.com.iotDevice.barchart
 
 import android.content.Context
+import android.widget.TextView
 import com.github.mikephil.charting.components.MarkerView
 import com.github.mikephil.charting.data.Entry
 import com.github.mikephil.charting.formatter.IAxisValueFormatter
 import com.github.mikephil.charting.highlight.Highlight
 import com.github.mikephil.charting.utils.MPPointF
 import iotdevice.com.iot_device.R
-import kotlinx.android.synthetic.main.custom_marker_view.view.*
 import java.text.DecimalFormat
 
-class XYMarkerView(context: Context, private val xAxisValueFormatter: IAxisValueFormatter, val enablePointDecimal: Boolean) : MarkerView(context, R.layout.custom_marker_view) {
+class XYMarkerView(context: Context, private val xAxisValueFormatter: ValueFormatter, val enablePointDecimal: Boolean) : MarkerView(context, R.layout.custom_marker_view) {
 
     private val format =  if (enablePointDecimal) {
         DecimalFormat("###.0")
@@ -25,7 +25,8 @@ class XYMarkerView(context: Context, private val xAxisValueFormatter: IAxisValue
         val tvText = xAxisValueFormatter.getFormattedValue(e!!.x, null) + ", " +
                 format.format(e.y.toDouble())
 
-        tvContent.text = tvText
+        findViewById<TextView>(R.id.tvContent).text = tvText
+//        tvContent.text = tvText
 
         super.refreshContent(e, highlight)
     }
