@@ -1,19 +1,18 @@
 package iotdevice.com.iotDevice.chart
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
 import android.content.res.Resources
+import android.util.Log
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.strongloop.android.loopback.callbacks.ListCallback
 import iotdevice.com.iotDevice.App
 import iotdevice.com.iotDevice.common.ChartUtils.Companion.calHourAndMinutes
 import iotdevice.com.iotDevice.model.DeviceStatusModel
 import iotdevice.com.iotDevice.repository.DeviceStatusRepository
 import iotdevice.com.iot_device.R
-import org.jetbrains.anko.AnkoLogger
-import org.jetbrains.anko.info
 import java.util.*
 
-class ChartViewModel : ViewModel(), AnkoLogger {
+class ChartViewModel : ViewModel() {
 
     val errorGetChart: MutableLiveData<Any> = MutableLiveData()
 
@@ -34,32 +33,91 @@ class ChartViewModel : ViewModel(), AnkoLogger {
         repository?.findTodayStatus(deviceId, object : ListCallback<DeviceStatusModel> {
             override fun onSuccess(objects: MutableList<DeviceStatusModel>?) {
 
-                val todayStatus = objects!![0]
+                // Fix a bug, device status will be updated during 0759 - 0800
+                // Server will create another device status record at that time,
+                // We need to handle the situation
+//                var productivity0Total = 0L
+//                var productivity1Total = 0L
+//                var productivity2Total = 0L
+//                var productivity3Total = 0L
+//                var productivity4Total = 0L
+//                var productivity5Total = 0L
+//                var productivity6Total = 0L
+//                var productivity7Total = 0L
+//                var productivity8Total = 0L
+//                var productivity9Total = 0L
+//                var productivity10Total = 0L
+//                var productivity11Total = 0L
+//                var productivity12Total = 0L
+//                var productivity13Total = 0L
+//                var productivity14Total = 0L
+//                var productivity15Total = 0L
+//                var productivity16Total = 0L
+//                var productivity17Total = 0L
+//                var productivity18Total = 0L
+//                var productivity19Total = 0L
+//                var productivity20Total = 0L
+//                var productivity21Total = 0L
+//                var productivity22Total = 0L
+//                var productivity23Total = 0L
 
-                totalOfDay.postValue(todayStatus.productivity0 +
-                        todayStatus.productivity1 +
-                        todayStatus.productivity2 +
-                        todayStatus.productivity3 +
-                        todayStatus.productivity4 +
-                        todayStatus.productivity5 +
-                        todayStatus.productivity6 +
-                        todayStatus.productivity7 +
-                        todayStatus.productivity8 +
-                        todayStatus.productivity9 +
-                        todayStatus.productivity10 +
-                        todayStatus.productivity11 +
-                        todayStatus.productivity12 +
-                        todayStatus.productivity13 +
-                        todayStatus.productivity14 +
-                        todayStatus.productivity15 +
-                        todayStatus.productivity16 +
-                        todayStatus.productivity17 +
-                        todayStatus.productivity18 +
-                        todayStatus.productivity19 +
-                        todayStatus.productivity20 +
-                        todayStatus.productivity21 +
-                        todayStatus.productivity22 +
-                        todayStatus.productivity23)
+
+
+//                objects?.forEach {
+//                    productivity0Total += it.productivity0
+//                    productivity1Total += it.productivity1
+//                    productivity2Total += it.productivity2
+//                    productivity3Total += it.productivity3
+//                    productivity4Total += it.productivity4
+//                    productivity5Total += it.productivity5
+//                    productivity6Total += it.productivity6
+//                    productivity7Total += it.productivity7
+//                    productivity8Total += it.productivity8
+//                    productivity9Total += it.productivity9
+//                    productivity10Total += it.productivity10
+//                    productivity11Total += it.productivity11
+//                    productivity12Total += it.productivity12
+//                    productivity13Total += it.productivity13
+//                    productivity14Total += it.productivity14
+//                    productivity15Total += it.productivity15
+//                    productivity16Total += it.productivity16
+//                    productivity17Total += it.productivity17
+//                    productivity18Total += it.productivity18
+//                    productivity19Total += it.productivity19
+//                    productivity20Total += it.productivity20
+//                    productivity21Total += it.productivity21
+//                    productivity22Total += it.productivity22
+//                    productivity23Total += it.productivity23
+//                }
+
+                objects?.last()?.let {
+                    totalOfDay.postValue(it.productivity0 +
+                            it.productivity1 +
+                            it.productivity2 +
+                            it.productivity3 +
+                            it.productivity4 +
+                            it.productivity5 +
+                            it.productivity6 +
+                            it.productivity7 +
+                            it.productivity8 +
+                            it.productivity9 +
+                            it.productivity10 +
+                            it.productivity11 +
+                            it.productivity12 +
+                            it.productivity13 +
+                            it.productivity14 +
+                            it.productivity15 +
+                            it.productivity16 +
+                            it.productivity17 +
+                            it.productivity18 +
+                            it.productivity19 +
+                            it.productivity20 +
+                            it.productivity21 +
+                            it.productivity22 +
+                            it.productivity23)
+                }
+
+
             }
 
             override fun onError(t: Throwable?) {
@@ -105,7 +163,12 @@ class ChartViewModel : ViewModel(), AnkoLogger {
                     val calendar = Calendar.getInstance()
                     calendar.timeInMillis = it.timeStamp!!.toLong()
 
-                    totalMonth += dayTotal
+                    // Server will save record within one day which caused by using UTC+0, different timezone between local and server
+                    dayTotalList[calendar.get(Calendar.DAY_OF_MONTH)] =  dayTotal.toFloat()
+                }
+
+                dayTotalList.forEach {
+                    totalMonth += it.value.toLong()
                 }
 
                 totalOfMonth.postValue(totalMonth)
@@ -246,7 +309,7 @@ class ChartViewModel : ViewModel(), AnkoLogger {
             }
 
             override fun onError(t: Throwable?) {
-                info("error : $t")
+                Log.i(tag, "error : $t")
                 headerItemLiveData.value = ChartHeaderItem(
                         calHourAndMinutes(0),
                         calHourAndMinutes(0),
@@ -255,5 +318,9 @@ class ChartViewModel : ViewModel(), AnkoLogger {
                 errorGetChart.value = t
             }
         })
+    }
+
+    companion object {
+        const val tag = "ChartViewModel"
     }
 }
