@@ -1,37 +1,39 @@
 package iotdevice.com.iotDevice.chart
 
 import android.graphics.Rect
-import android.support.v7.widget.RecyclerView
+
 import android.view.View
+import androidx.recyclerview.widget.RecyclerView
 
 class GridLayoutDivider(val spanCount: Int, val spacing: Int, val includeEdge: Boolean, val headerNum: Int) : RecyclerView.ItemDecoration() {
-    override fun getItemOffsets(outRect: Rect?, view: View?, parent: RecyclerView?, state: RecyclerView.State?) {
-        val position = parent!!.getChildAdapterPosition(view) - headerNum // item position
+
+    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+        val position = parent.getChildAdapterPosition(view) - headerNum // item position
 
         if (position >= 0) {
             val column = position % spanCount // item column
 
             if (includeEdge) {
-                outRect?.left = spacing - column * spacing / spanCount // spacing - column * ((1f / spanCount) * spacing)
-                outRect?.right = (column + 1) * spacing / spanCount // (column + 1) * ((1f / spanCount) * spacing)
+                outRect.left = spacing - column * spacing / spanCount // spacing - column * ((1f / spanCount) * spacing)
+                outRect.right = (column + 1) * spacing / spanCount // (column + 1) * ((1f / spanCount) * spacing)
 
                 if (position < spanCount) { // top edge
-                    outRect?.top = spacing
+                    outRect.top = spacing
                 }
-                outRect?.bottom = spacing // item bottom
+                outRect.bottom = spacing // item bottom
             } else {
-                outRect?.left = column * spacing / spanCount // column * ((1f / spanCount) * spacing)
-                outRect?.right = spacing - (column + 1) * spacing / spanCount // spacing - (column + 1) * ((1f /    spanCount) * spacing)
+                outRect.left = column * spacing / spanCount // column * ((1f / spanCount) * spacing)
+                outRect.right = spacing - (column + 1) * spacing / spanCount // spacing - (column + 1) * ((1f /    spanCount) * spacing)
                 if (position >= spanCount) {
-                    outRect?.top = spacing // item top
+                    outRect.top = spacing // item top
                 }
             }
         } else {
-            outRect?.left = spacing
-            outRect?.right = spacing
-            outRect?.top = spacing
+            outRect.left = spacing
+            outRect.right = spacing
+            outRect.top = spacing
             // Item will get top spacing
-            outRect?.bottom = 0
+            outRect.bottom = 0
         }
     }
 }
