@@ -1,7 +1,8 @@
 package iotdevice.com.iotDevice.common
 
+import android.app.AlertDialog
 import android.content.Context
-import android.support.v7.app.AlertDialog
+
 import iotdevice.com.iot_device.R
 
 class DialogUtils {
