@@ -1,10 +1,11 @@
 package iotdevice.com.iotDevice.splash
 
+import android.content.Intent
 import android.os.Bundle
 import iotdevice.com.iotDevice.common.BaseActivity
 import iotdevice.com.iotDevice.login.LoginActivity
 import iotdevice.com.iot_device.R
-import org.jetbrains.anko.startActivity
+
 import java.util.*
 import kotlin.concurrent.schedule
 
@@ -15,7 +16,8 @@ class SplashActivity: BaseActivity() {
         setContentView(R.layout.activity_splash)
 
         Timer().schedule(1000){
-            startActivity<LoginActivity>()
+            val intent = Intent(applicationContext, LoginActivity::class.java)
+            startActivity(intent)
             finish()
         }
     }
